Extract duplicate test code into separate module
