+++ near-sdk/src/test_utils/mod.rs
@@ -1,0 +1,54 @@
+pub(crate) mod test_env {
+    use crate::collections::Set;
+    use crate::{env, MockedBlockchain};
+    use near_vm_logic::types::AccountId;
+    use near_vm_logic::VMContext;
+    use rand::seq::SliceRandom;
+    use rand::{Rng, SeedableRng};
+    use std::collections::HashSet;
+    use std::iter::FromIterator;
+
+    fn alice() -> AccountId {
+        "alice.near".to_string()
+    }
+
+    fn bob() -> AccountId {
+        "bob.near".to_string()
+    }
+
+    fn carol() -> AccountId {
+        "carol.near".to_string()
+    }
+
+    pub(crate) fn setup() {
+        let context = VMContext {
+            current_account_id: alice(),
+            signer_account_id: bob(),
+            signer_account_pk: vec![0, 1, 2],
+            predecessor_account_id: carol(),
+            input: vec![],
+            block_index: 0,
+            block_timestamp: 0,
+            account_balance: 0,
+            account_locked_balance: 0,
+            storage_usage: 10u64.pow(6),
+            attached_deposit: 0,
+            prepaid_gas: 10u64.pow(18),
+            random_seed: vec![0, 1, 2],
+            is_view: false,
+            output_data_receivers: vec![],
+            epoch_height: 0,
+        };
+        let storage = match env::take_blockchain_interface() {
+            Some(mut bi) => bi.as_mut_mocked_blockchain().unwrap().take_storage(),
+            None => Default::default(),
+        };
+        env::set_blockchain_interface(Box::new(MockedBlockchain::new(
+            context,
+            Default::default(),
+            Default::default(),
+            vec![],
+            storage,
+        )));
+    }
+}
